+++ ficary/__init__.py
@@ -2,7 +2,7 @@
 
 import logging as _logging
 
-__version__ = "2.18.4"
+__version__ = "2.19.0"
 
 _logger = _logging.getLogger(__name__)